Implement stopping targeting area on user click
The coordinates of the clicked location are logged in the console as a
ratio of the width and height of the image. This ratio will be used to
store the coordinates of the characters in the image since the size of
the image can change depending on the size of the user's browser
window.