Implement element overlapping checks
The element overlapping check logic is needed in order to see whether
the user hovered the targeting box over another element that
represents a character in the image.

For now, this is done by using the bounding client rect for the
targeting box element and the character box elements. The DOMRect
objects are passed into the checkOverlap function which returns true
or false depending on whether the two elements are overlapping or not.

This will be done on the backend by storing the values of left, right,
top, and bottom from each of the character box elements' DOMRect
objects.