Finish implementation of high score backend
The app now properly prompts the user to input their name if the
length of the list of high scores is less than 10 or if the user has a
faster time than the slowest time in the list of high scores.

The app will not prompt the user if their time is not faster than the
top 10 times on the high score list in the backend.

The results menu still needs to be styled and the user's name needs to
be validated to have only 3 characters.