Finish styling results menu
The form data for the user's name on the leaderboard is transformed to
all uppercase before sending the data to the firestore backend.

The input element that allows the user to submit their username was
also made to validate the user input so it will only accept names of
length 3 alphanumeric.